+++ pkg/storegateway/bucket_index_postings_test.go
@@ -36,7 +36,7 @@ func TestBigEndianPostingsCount(t *testing.T) {
 	assert.Equal(t, count, c)
 }
 
-func TestWorstCaseFetchedDataStrategy(t *testing.T) {
+func TestMinimizeFetchedDataStrategy(t *testing.T) {
 	testCases := map[string]struct {
 		input            []postingGroup
 		expectedSelected []postingGroup
@@ -116,116 +116,7 @@ func TestWorstCaseFetchedDataStrategy(t *testing.T) {
 
 	for testName, testCase := range testCases {
 		t.Run(testName, func(t *testing.T) {
-			actualSelected, actualOmitted := worstCaseFetchedDataStrategy{1.0}.selectPostings(testCase.input)
-			assert.ElementsMatch(t, testCase.expectedSelected, actualSelected)
-			assert.ElementsMatch(t, testCase.expectedOmitted, actualOmitted)
-		})
-	}
-}
-
-func TestSpeculativeFetchedDataStrategy(t *testing.T) {
-	testCases := map[string]struct {
-		input            []postingGroup
-		expectedSelected []postingGroup
-		expectedOmitted  []postingGroup
-	}{
-		"single posting group is selected": {
-			input: []postingGroup{
-				{totalSize: 128},
-			},
-			expectedSelected: []postingGroup{
-				{totalSize: 128},
-			},
-		},
-		"only all-postings & subtracting groups": {
-			input: []postingGroup{
-				{totalSize: 0 /* all-postings doesn't have a size at the moment */, keys: []labels.Label{allPostingsKey}},
-				{isSubtract: true, totalSize: 128},
-				{isSubtract: true, totalSize: 64 * 1024 * 1024},
-			},
-			expectedSelected: []postingGroup{
-				{totalSize: 0, keys: []labels.Label{allPostingsKey}},
-				{isSubtract: true, totalSize: 128},
-				{isSubtract: true, totalSize: 64 * 1024 * 1024},
-			},
-		},
-		"only small posting lists": {
-			input: []postingGroup{
-				{totalSize: 1024},
-				{totalSize: 256},
-				{totalSize: 128},
-			},
-			expectedSelected: []postingGroup{
-				{totalSize: 1024},
-				{totalSize: 256},
-				{totalSize: 128},
-			},
-		},
-		"two small and one large list": {
-			input: []postingGroup{
-				{totalSize: 64 * 1024 * 1024},
-				{totalSize: 256},
-				{totalSize: 128},
-			},
-			expectedSelected: []postingGroup{
-				{totalSize: 256},
-				{totalSize: 128},
-			},
-			expectedOmitted: []postingGroup{
-				{totalSize: 64 * 1024 * 1024},
-			},
-		},
-		"one small and two mid size lists - last is omitted (more aggressive than worstCase)": {
-			input: []postingGroup{
-				{totalSize: 128},
-				{totalSize: 4 * 1024},
-				{totalSize: 4 * 1024},
-			},
-			expectedSelected: []postingGroup{
-				{totalSize: 128},
-				{totalSize: 4 * 1024},
-			},
-			expectedOmitted: []postingGroup{
-				{totalSize: 4 * 1024},
-			},
-		},
-		"subtractive lists are not taken into consideration for halving number of postings": {
-			input: []postingGroup{
-				{totalSize: 128},
-				{isSubtract: true, totalSize: 256},
-				{isSubtract: true, totalSize: 512},
-				{totalSize: 4 * 1024},
-				{totalSize: 4 * 1024},
-			},
-			expectedSelected: []postingGroup{
-				{totalSize: 128},
-				{isSubtract: true, totalSize: 256},
-				{isSubtract: true, totalSize: 512},
-				{totalSize: 4 * 1024},
-			},
-			expectedOmitted: []postingGroup{
-				{totalSize: 4 * 1024},
-			},
-		},
-		"x": {
-			input: []postingGroup{
-				{totalSize: 4 * 1024 * 1024},
-				{totalSize: 32*1024 + 4},
-				{totalSize: 32*1024 + 4},
-			},
-			expectedSelected: []postingGroup{
-				{totalSize: 32*1024 + 4},
-				{totalSize: 32*1024 + 4},
-			},
-			expectedOmitted: []postingGroup{
-				{totalSize: 4 * 1024 * 1024},
-			},
-		},
-	}
-
-	for testName, testCase := range testCases {
-		t.Run(testName, func(t *testing.T) {
-			actualSelected, actualOmitted := speculativeFetchedDataStrategy{}.selectPostings(testCase.input)
+			actualSelected, actualOmitted := minimizeFetchedDataStrategy{}.selectPostings(testCase.input)
 			assert.ElementsMatch(t, testCase.expectedSelected, actualSelected)
 			assert.ElementsMatch(t, testCase.expectedOmitted, actualOmitted)
 		})
@@ -297,7 +188,7 @@ func TestLabelValuesPostingsStrategy(t *testing.T) {
 	for testName, testCase := range testCases {
 		t.Run(testName, func(t *testing.T) {
 			strategy := labelValuesPostingsStrategy{
-				matchersStrategy: worstCaseFetchedDataStrategy{1},
+				matchersStrategy: minimizeFetchedDataStrategy{},
 				allLabelValues:   testCase.postingLists,
 			}
 			actualSelected, actualOmitted := strategy.selectPostings(testCase.input)